Move shouldLogErrorAndReturnBadGatewayWhenInterruptionOccurs to FetchByIdentifierAndYearHandlerTest. Move common tests for journal/series to FetchSerialPublicationByIdentifierAndYearHandlerTest
